feat: Add healthcheck script and documentation

+++ Dockerfile
@@ -49,8 +49,10 @@ RUN bash ${SALT_BUILD_DIR}/install.sh
 COPY assets/runtime ${SALT_RUNTIME_DIR}
 RUN chmod -R +x ${SALT_RUNTIME_DIR}
 
+COPY assets/bin/* /usr/local/bin
+
 # Cleaning tasks
-RUN rm -rf ${SALT_BUILD_DIR}/*
+RUN rm -rf "${SALT_BUILD_DIR:?}"/*
 
 # Entrypoint
 COPY entrypoint.sh /sbin/entrypoint.sh
